requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
asgiref==3.4.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
cycler==0.10.0
fastapi==0.68.1
fire==0.4.0
flatbuffers==1.12
gast==0.4.0
gevent==21.1.2
google-auth==1.28.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.0.0
grequests==0.6.0
grpcio==1.34.1
h11==0.12.0
h5py==3.1.0
idna==2.10
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.1
multidict==5.1.0
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
pandas==1.2.4
Pillow==8.3.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-pretrained-vit==0.0.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.2
seaborn==0.11.1
six==1.15.0
starlette==0.14.2
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0rc0
termcolor==1.1.0
tf-estimator-nightly==2.5.0.dev2021032501
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.15.0
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.3.0
